+++ src/main/apiClient.ts
@@ -104,6 +104,16 @@ export type TokenRequestParameters = {
   persistent?: boolean // Whether the remember me is enabled
 }
 
+export type StartPhoneNumberRegistrationParams = {
+  accessToken: string
+  phoneNumber: string
+}
+
+export type VerifyPhoneNumberRegistrationParams = {
+  accessToken: string
+  verificationCode: string
+}
+
 type AuthenticationToken = { tkn: string }
 
 /**
@@ -123,6 +133,8 @@ export default class ApiClient {
     this.authorizeUrl = `https://${this.config.domain}/oauth/authorize`
     this.tokenUrl = `https://${this.config.domain}/oauth/token`
     this.popupRelayUrl = `https://${this.config.domain}/popup/relay`
+    this.startPhoneNumberRegistrationUrl = `https://${this.baseUrl}/mfa/credentials/register/phone-number/start`
+    this.verifyPhoneNumberRegistrationUrl = `https://${this.baseUrl}/mfa/credentials/register/phone-number/verify`
 
     this.initCordovaCallbackIfNecessary()
   }
@@ -135,6 +147,8 @@ export default class ApiClient {
   private authorizeUrl: string
   private tokenUrl: string
   private popupRelayUrl: string
+  private readonly startPhoneNumberRegistrationUrl: string
+  private readonly verifyPhoneNumberRegistrationUrl: string
 
   getSignupData(signupToken: string): Promise<OpenIdUser> {
     return this.http.get<OpenIdUser>(`${this.baseUrl}/signup/data`, {
@@ -826,6 +840,26 @@ export default class ApiClient {
 
   removeWebAuthnDevice(accessToken: string, deviceId: string): Promise<void> {
     return this.http.remove<void>(`/webauthn/registration/${deviceId}`, { accessToken })
+  }
+
+  startPhoneNumberRegistration(params: StartPhoneNumberRegistrationParams): Promise<void> {
+    const { accessToken, phoneNumber } = params
+    return this.http.post<void>(this.startPhoneNumberRegistrationUrl, {
+      body: {
+        phoneNumber
+      },
+      accessToken
+    })
+  }
+
+  verifyPhoneNumberRegistration(params: VerifyPhoneNumberRegistrationParams): Promise<void> {
+    const { accessToken, verificationCode } = params
+    return this.http.post<void>(this.verifyPhoneNumberRegistrationUrl, {
+      body: {
+        verificationCode
+      },
+      accessToken
+    })
   }
 
   getSessionInfo(): Promise<SessionInfo> {
